[CL 13191154 by Marc Audy in Dev-Framework branch]

+++ Engine/Source/Runtime/AudioMixer/Private/SoundFileIO/SoundFileIOEnums.h
@@ -165,7 +165,8 @@ namespace Audio
 	 */
 	namespace ESoundFileChannelMap
 	{
-		enum class Type : uint8
+		// this is used to populate an array which is passed into a sound file API call so must be uint32
+		enum class Type : uint32
 		{
 			INVALID = 0,
 			MONO,
